lxd: Remove metrics functions from `request` package
`MetricsCallback` must be moved to avoid an import loop between the auth
and the request packages. Thus, `CountStartedRequest` is moved as well
to keep them together.
