+++ lxd/metrics/api_rates.go
@@ -1,10 +1,15 @@
 package metrics
 
 import (
+	"net/http"
 	"net/url"
+	"strings"
+	"sync"
 	"sync/atomic"
 
+	"github.com/canonical/lxd/lxd/request"
 	"github.com/canonical/lxd/shared/entity"
+	"github.com/canonical/lxd/shared/logger"
 )
 
 // RequestResult represents a completed request status category.
@@ -71,4 +76,35 @@ func GetOngoingRequests(entityType entity.Type) int64 {
 // GetCompletedRequests gets the value of completed requests filtered by entity type and result.
 func GetCompletedRequests(entityType entity.Type, result RequestResult) int64 {
 	return completedRequests[completedMetricsLabeling{entityType: entityType, result: result}].Load()
+}
+
+// CountStartedRequest tracks the request as started for the API metrics and
+// injects a callback function to track the request as completed.
+func CountStartedRequest(r *http.Request) {
+	requestURL := *r.URL
+
+	// Set the callback function to track the request as completed.
+	// Use sync.Once to ensure it can be called at most once.
+	var once sync.Once
+	callbackFunc := func(result RequestResult) {
+		once.Do(func() {
+			TrackCompletedRequest(requestURL, result)
+		})
+	}
+
+	request.SetCtxValue(r, request.MetricsCallbackFunc, callbackFunc)
+
+	TrackStartedRequest(requestURL)
+}
+
+// MetricsCallback retrieves a callback function from the request context and calls it.
+// The callback function is used to mark the request as completed for the API metrics.
+func MetricsCallback(req *http.Request, result RequestResult) {
+	callback, err := request.GetCtxValue[func(RequestResult)](req.Context(), request.MetricsCallbackFunc)
+	if err != nil && (strings.HasPrefix(req.URL.Path, "/1.0") || req.URL.Path == "/") {
+		// Log a warning if endpoint is part of the main API, and therefore should be counted fot the API metrics.
+		logger.Warn("Request will not be counted for the API metrics", logger.Ctx{"url": req.URL.Path, "method": req.Method, "remote": req.RemoteAddr})
+	} else if err == nil && callback != nil {
+		callback(result)
+	}
 }
